Refactor, add docs.
